refactor: rename events filename to PascalCase
