--- src/Events/Client/w.js
@@ -1,9 +1,0 @@
-const Event = require('../../Structures/Event');
-
-module.exports = class extends Event {
-
-	async run(info) {
-		this.client.logger.warn(info);
-	}
-
-};
