+++ src/Events/ClientReady.js
@@ -1,17 +1,18 @@
 const Event = require('../Structures/Event');
 const { ActivityType } = require('discord-api-types/v10');
-const colorette = require('colorette');
+const Colorette = require('colorette');
 
 module.exports = class extends Event {
 
 	constructor(...args) {
 		super(...args, {
+			name: 'ready',
 			once: true
 		});
 	}
 
 	async run() {
-		this.client.logger.log(`Logged in as ${colorette.redBright(`${this.client.user.tag}`)}`, { status: 'BOOT', color: 'greenBright' });
+		this.client.logger.log(`Logged in as ${Colorette.redBright(`${this.client.user.tag}`)}`, { status: 'READY', color: 'greenBright' });
 		this.client.logger.log(`Loaded ${(this.client.commands.size + this.client.interactions.size).formatNumber()} commands & ${this.client.events.size.formatNumber()} events!`);
 
 		const activities = [
