+++ src/Events/Guild/GuildCreate.js
@@ -5,21 +5,30 @@ const { Access, Colors } = require('../../Utils/Constants');
 
 module.exports = class extends Event {
 
+	constructor(...args) {
+		super(...args, {
+			name: 'guildCreate',
+			once: false
+		});
+	}
+
 	async run(guild) {
 		if (!guild.available) return;
 
+		if (!Access.GuildLogWebhook) return;
 		const webhook = new WebhookClient({ url: Access.GuildLogWebhook });
-		if (!webhook) return;
 
 		const guildOwner = await guild.fetchOwner();
 
 		const embed = new EmbedBuilder()
-			.setColor(Colors.Red)
-			.setTitle(`${this.client.user.username} left a Server!`)
+			.setColor(Colors.Green)
+			.setTitle(`${this.client.user.username} was added to a new Server!`)
 			.setThumbnail(guild.iconURL() ? guild.iconURL({ size: 512 }) : `https://guild-default-icon.herokuapp.com/${encodeURIComponent(guild.nameAcronym)}`)
 			.setDescription([
 				`***Server:*** ${guild.name} (\`${guild.id}\`)`,
-				`***Owner:*** ${guildOwner.user.tag} (\`${guildOwner.id}\`)`
+				`***Owner:*** ${guildOwner.user.tag} (\`${guildOwner.id}\`)`,
+				`***Channels:*** ${guild.channels.cache.size.formatNumber()}`,
+				`***Members:*** ${guild.memberCount.formatNumber()}`
 			].join('\n'))
 			.setFooter({ text: `${this.client.guilds.cache.size.formatNumber()} guilds | ${this.client.guilds.cache.reduce((a, b) => a + b.memberCount, 0).formatNumber()} users`, iconURL: this.client.user.avatarURL() });
