+++ src/Events/Node/UnhandledRejection.js
@@ -1,15 +1,17 @@
 const Event = require('../../Structures/Event');
+const process = require('node:process');
 
 module.exports = class extends Event {
 
 	constructor(...args) {
 		super(...args, {
+			name: 'unhandledRejection',
+			once: false,
 			emitter: process
 		});
 	}
 
-	// eslint-disable-next-line no-unused-vars
-	async run(error, promise) {
+	async run(error, promise) { // eslint-disable-line no-unused-vars
 		this.client.logger.error(error.stack);
 	}
